+++ apps/api/app/repositories/crawl_stats_repository.py
@@ -1,8 +1,8 @@
-from datetime import date
+from datetime import date, timedelta
 from typing import List, Optional
 
 from sqlalchemy import select
-from sqlalchemy.orm import joinedload
+from sqlalchemy.orm import joinedload, selectinload
 
 from app.core.db import async_session
 from app.core.logger import get_logger
@@ -51,3 +51,15 @@ class CrawlStatsRepository:
             except Exception as e:
                 logger.error(f"Failed to get crawl stats by date range: {e}")
                 return []
+
+    @staticmethod
+    async def get_crawl_stats_last_day() -> list[CrawlStats]:
+        async with async_session() as session:
+            today = date.today()
+            yesterday = today - timedelta(days=1)
+            result = await session.execute(
+                sa.select(CrawlStats)
+                .where(CrawlStats.crawl_date >= yesterday)
+                .options(selectinload(CrawlStats.subscription))
+            )
+            return result.scalars().all()
